requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
python-multipart==0.0.5
google-api-python-client==2.86.0
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
pytubefix
moviepy==1.0.3
python-dotenv==0.19.0
pandas>=2.0.0
numpy>=1.23.0
scikit-learn
matplotlib
seaborn==0.11.2
certifi>=2024.2.2
google-cloud-storage
